+++ Rdmp.Dicom/Extraction/FoDicomBased/FoDicomAnonymiser.cs
@@ -57,6 +57,9 @@ namespace Rdmp.Dicom.Extraction.FoDicomBased
         [DemandsInitialization("Set to true to skip anonymisation process on structured reports (Modality=SR).  PatientID and UID tags will still be anonymised.", DefaultValue = false)]
         public bool SkipAnonymisationOnStructuredReports { get; set; }
 
+        [DemandsInitialization("Set to true to skip opening/anonymising files and just process the metadata already in the database.", DefaultValue = false)]
+        public bool MetadataOnly { get; set; }
+
         private IPutDicomFilesInExtractionDirectories _putter;
 
         private int _anonymisedImagesCount = 0;
@@ -82,6 +85,28 @@ namespace Rdmp.Dicom.Extraction.FoDicomBased
                 return toProcess;
             }
             
+            if(MetadataOnly)
+            {
+                var matching = toProcess.Columns.Cast<DataColumn>().Where(
+                    c => UIDMapping.SupportedTags.Any(k => k.Key.DictionaryEntry.Keyword.Equals(c.ColumnName)))
+                    .ToArray();
+                
+                if (!matching.Any())
+                {
+                    listener.OnNotify(this, new(ProgressEventType.Information, "Ignoring non imaging dataset, it had no UID columns"));
+                    return toProcess;
+                }
+
+                Dictionary<DataColumn, KeyValuePair<DicomTag, UIDType>> dictionary = matching.ToDictionary(k => k,
+                        c => UIDMapping.SupportedTags.First(k => k.Key.DictionaryEntry.Keyword.Equals(c.ColumnName)));
+
+                foreach (DataRow row in toProcess.Rows)
+                {
+                    SubstituteMetadataOnly(row, dictionary);
+                }
+                return toProcess;
+            }
+
             //if it isn't a dicom dataset don't process it
             if (!toProcess.Columns.Contains(RelativeArchiveColumnName))
             {
@@ -123,6 +148,23 @@ namespace Rdmp.Dicom.Extraction.FoDicomBased
             _sw.Stop();
 
             return toProcess;
+        }
+
+        private void SubstituteMetadataOnly(DataRow row, Dictionary<DataColumn, KeyValuePair<DicomTag, UIDType>> dictionary)
+        {
+            //rewrite the UIDs
+            foreach (var kvp in dictionary)
+            {
+                // no UID substitution server so no UID subs
+                if (_uidSubstitutionLookup == null)
+                    throw new Exception($"{nameof(MetadataOnly)} is on but there is no UID lookup server configured");
+
+                var value = row[kvp.Key].ToString();
+
+                //if it has a value for this UID
+                if (value == null) continue;
+                row[kvp.Key] = _uidSubstitutionLookup.GetOrAllocateMapping(value, _projectNumber, kvp.Value.Value);
+            }
         }
 
         /// <summary>